requirements.txt
aiofiles==22.1.0
ajsonrpc==1.2.0
anyio==3.6.2
bottle==0.12.23
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
h11==0.14.0
idna==3.4
marshmallow==3.19.0
packaging==23.0
paho-mqtt==1.6.1
platformio==6.1.6
psycopg2-binary==2.9.5
pyelftools==0.29
pyserial==3.5
requests==2.28.2
semantic-version==2.10.0
sniffio==1.3.0
starlette==0.23.1
tabulate==0.9.0
urllib3==1.26.14
uvicorn==0.20.0
wsproto==1.2.0
